Move sales engine mocks to data directory
